refactor(domain): DRY and modularize the on-chain (HIP-5) support
Follow-up to the Kody-flagged review. No behavior change; extracted shared
pieces and split the on-chain logic into its own file:

- Moves ConvertToOnChain and its sentinel errors into internal/service/domain/
  onchain.go, matching the established per-concern file split
  (platform_domain.go), shrinking delegated_domain_service.go.
- Extracts DelegatedDomainService.assignPrimaryAndNotify, replacing the
  identical set-primary + fire-created-notification block triplicated across
  the managed, self-hosted, and on-chain-managed bind paths.
- Extracts api.parseDomainPath + loadDomainBinding, replacing the duplicated
  auth + :id/:domain_id parsing + not-found handling across
  domainDNSRequirements, republishDomainDANE, and convertDomainToOnChain.
- Extracts website.onchainDNSHostingUnavailableError, shared by the
  SetDomainDNSEnabled and UpdateWebsite on-chain enable refusals.
